up
up

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: leith <leith@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/01/21 17:22:19 by leith             #+#    #+#             */
-/*   Updated: 2017/03/07 12:49:58 by aazri            ###   ########.fr       */
+/*   Updated: 2017/03/09 11:25:23 by aazri            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,8 +73,8 @@ size_t	ft_nbulen(unsigned long n, unsigned int base);
 size_t	ft_nblen(long n, unsigned int base);
 void	ft_putnstr(char *s, size_t max);
 void	print_base(uintmax_t nb, unsigned int base);
-size_t adapt_width(t_flags *flags, char *sign, size_t precision, uintmax_t nb, size_t *nb_len);
-size_t adapt_precision(t_flags *flags, size_t nb_len);
+unsigned int adapt_width(t_flags *flags, char *sign, size_t precision, uintmax_t nb, size_t *nb_len);
+unsigned int adapt_precision(t_flags *flags, size_t nb_len);
 int print_count(size_t nb_len, size_t pad_len, t_flags *flags, char *prefix, uintmax_t nb);
 char *base_convert(uintmax_t nb, unsigned int base);
